Redux setup
Co-Authored by Edwin Del Bosque <edwindelbosque@gmail.com>
