+++ actions/index.js
@@ -1,0 +1,4 @@
+export const setSelect = data => ({
+  type: 'OBJECT',
+  data
+});
