+++ package.json
@@ -18,7 +18,8 @@
     "react-navigation-stack": "^1.10.3",
     "react-redux": "^7.1.3",
     "redux": "^4.0.4",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "redux-logger": "^3.0.6"
   },
   "devDependencies": {
     "babel-preset-expo": "~8.0.0"
