+++ reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import select from './select';
+
+const rootReducer = combineReducers({
+  select
+  // upcomingMovies,
+
+});
+
+export default rootReducer;
