+++ reducers/select.js
@@ -1,0 +1,10 @@
+const select = (state = [], action) => {
+  switch (action.type) {
+    case 'OBJECT':
+      return action.favorites;
+    default:
+      return state;
+  }
+};
+
+export default select;
